# docs/04-技术架构/当前版本/V1-openapi.yaml
openapi: 3.0.3
info:
  title: DGOS V1 Logical API
  version: "1.0.0"
  description: |
    DGOS V1 逻辑契约的 HTTPS JSON + SSE 投影。桌面 IPC 必须保持相同的动作、字段、权限和事件语义。
    该文件同时登记已实现的本地 contract projection 与尚未具备真实桌面/Web E2E 证据的 draft operation。
security:
  - dgosSession: []
  - dgosApiKey: []
servers:
  - url: https://localhost
    description: Docker Compose Web projection (planned)
...
paths:
...
  /api/v1/actions/{actionId}/plan:
    post:
      tags: [assistant]
      operationId: planAction
      x-tag-feature: V1-FR-009
      x-tag-implementation-status: implemented-local-contract
      parameters:
        - { $ref: '#/components/parameters/ActionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionPlanRequest' }
      responses:
        '200':
          description: Action plan and confirmation requirement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionPlan' }
  /api/v1/actions/{actionId}/execute:
    post:
      tags: [assistant]
      operationId: executeAction
      x-tag-feature: V1-FR-009
      x-tag-implementation-status: implemented-local-contract
      parameters:
        - { $ref: '#/components/parameters/ActionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionExecuteRequest' }
      responses:
        '202':
          description: Action execution accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRun' }
  /api/v1/action-runs/{runId}:
    get:
      tags: [assistant]
      operationId: getActionRun
      x-tag-feature: V1-FR-009
      x-tag-implementation-status: implemented-local-contract
      parameters:
        - { $ref: '#/components/parameters/ActionRunId' }
      responses:
        '200':
          description: Action run snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRun' }
    delete:
      tags: [assistant]
      operationId: cancelActionRun
      x-tag-feature: V1-FR-009
      x-tag-implementation-status: implemented-local-contract
      parameters:
        - { $ref: '#/components/parameters/ActionRunId' }
      responses:
        '202':
          description: Cancellation accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRun' }
...
components:
...
  securitySchemes:
    appLaunchTicket:
      type: apiKey
      in: query
      name: launchTicket
      description: Opaque short-lived session/subject/app/package/instance-bound static ticket, redacted from logs; no bridge authority.
    dgosSession:
      type: apiKey
      in: cookie
      name: dgos_session
      description: Revocable DGOS administrator or subject session cookie.
    dgosApiKey:
      type: http
      scheme: bearer
      bearerFormat: DGOS-API-Key
      description: DGOS API Key; the secret is never returned after creation.
...
  parameters:
...
    ActionId: { in: path, name: actionId, required: true, schema: { type: string } }
    ActionRunId: { in: path, name: runId, required: true, schema: { type: string } }
...
  schemas:
...
    ActionNavigation:
      type: object
      additionalProperties: false
      required: [target]
      properties:
        target: { type: string, enum: [system.settings, provider.settings, skill.management, mcp.management, app.catalog] }
...
    RequestEnvelope:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string }
...
    ActionPlanRequest:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [input]
          properties:
            input: { type: object, additionalProperties: true }
    ActionPlan:
      type: object
      required: [actionId, actionVersion, input, confirmation, risk]
      properties:
        actionId: { type: string }
        actionVersion: { type: string }
        input: { type: object, additionalProperties: true }
        confirmation: { type: string, enum: [none, required, elevated] }
        risk: { type: string, enum: [read, write, external, destructive] }
        targetSummary: { type: string }
        inputSummary: { type: string }
    ActionExecuteRequest:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [input]
          properties:
            input: { type: object, additionalProperties: true }
            confirmationToken: { type: string, writeOnly: true }
    ActionRun:
      type: object
      required: [runId, actionId, actionVersion, requestId, status]
      properties:
        runId: { type: string }
        actionId: { type: string }
        actionVersion: { type: string }
        requestId: { type: string }
        taskId: { type: string }
        status: { type: string, enum: [planned, awaiting-confirmation, authorized, running, succeeded, denied, cancelled, failed, expired, missing] }
        result: { type: object, additionalProperties: true }
        resultSummary:
          type: object
          properties:
            navigation: { $ref: '#/components/schemas/ActionNavigation' }
        error: { type: object, additionalProperties: true }
...